refactor: linking for headlines
Previously linking to a headline did not work correctly, for multiple
reasons. First, Typst does not understand two labels for a headline. We now
assume that an element within a headline, wants to link to that
headline. Therefore, we always link to the headline, instead of creating a label
inside the headline. Another reason for this, is Typst again. Typst will not
filter labels out of headlines, when creating a TOC. As a result, we have
duplicate labels with the same name within the document.

This newly implemented behavior is the most usable and expectable. However, it
might not be the correct behavior.