refactor: remove oc-typst.el
This file was always part of `ox-typst.el`. Some linting tools are not happy
with the name and their names. The previous layout was mirror from the real Org
repository. There is no difference for us doing it this way and it pleases the
linter. There is no reason to use the Typst citation exporter without
`ox-typst`.